Making strides
Currently as follows: Checks if user exists. If the user does NOT exist,
it will register them, then move into... If the user exists, pulls all
the info for them, IN ADDITION TO the question they are on and all the
possible answers associated with it.

It only prints to console for now, but need to extract the choices from
it. Possible course of action: Do a second query JUST TO GET the
answers, then store those in variables, so they can be easily (and
individually) manipulated.